+++ siade/app/controllers/api_particulier/v3_and_more/cnav/allocation_rentree_scolaire_with_france_connect_controller.rb
@@ -1,0 +1,11 @@
+class APIParticulier::V3AndMore::CNAV::AllocationRentreeScolaireWithFranceConnectController < APIParticulier::V3AndMore::CNAV::AbstractFranceConnectController
+  private
+
+  def organizer_class
+    CNAV::AllocationRentreeScolaire
+  end
+
+  def serializer_module
+    ::APIParticulier::CNAV::AllocationRentreeScolaire
+  end
+end