+++ siade/app/interactors/cnav/allocation_rentree_scolaire/build_resource.rb
@@ -1,0 +1,10 @@
+class CNAV::AllocationRentreeScolaire::BuildResource < CNAV::BuildResource
+  protected
+
+  def resource_attributes
+    {
+      status:,
+      date_debut_droit:
+    }
+  end
+end