+++ siade/app/interactors/cnav/allocation_rentree_scolaire/validate_response.rb
@@ -1,0 +1,7 @@
+class CNAV::AllocationRentreeScolaire::ValidateResponse < CNAV::ValidateResponse
+  private
+
+  def valid_indicateurs
+    %w[ALLOCATAIRE OUVRANT_DROIT NON_BENEFICIAIRE]
+  end
+end