+++ siade/app/interactors/cnav/prime_activite/validate_response.rb
@@ -1,0 +1,7 @@
+class CNAV::PrimeActivite::ValidateResponse < CNAV::ValidateResponse
+  private
+
+  def valid_indicateurs
+    %w[BENEFICIAIRE NON_BENEFICIAIRE]
+  end
+end