+++ siade/app/interactors/cnav/revenu_solidarite_active/validate_response.rb
@@ -1,0 +1,7 @@
+class CNAV::RevenuSolidariteActive::ValidateResponse < CNAV::ValidateResponse
+  private
+
+  def valid_indicateurs
+    %w[BENEFICIAIRE NON_BENEFICIAIRE]
+  end
+end